renamed check-daemons-status file, removed some comments, added tabulations to downReason variable
