+++ public/controllers/settings/settings.js
@@ -189,7 +189,7 @@ export class SettingsController {
         } catch (error) {
           const code = ((error || {}).data || {}).code;
           const downReason = typeof error === 'string' ? error :
-          (error || {}).message || ((error || {}).data || {}).message || 'Wazuh is not reachable';
+            (error || {}).message || ((error || {}).data || {}).message || 'Wazuh is not reachable';
           const status = code === 3099 ? 'down' : 'unknown';
           this.apiEntries[idx].status = { status, downReason };
           numError = numError + 1;
